Support MSVC builds of PyGObject-3.x
The official compiler that is supported by Python is Visual Studio, namely
2008 for 2.7.x and 3.2.x, 2010 for 3.3.x and 3.4.x and 2015 for 3.5.x and
later, so this is work to add build support for Visual Studio via NMake,
so that we can reduce the problems that are caused by different CRTs that
are used by Python and PyGObject builds, as GLib and g-i supported builds
using Visual Studio 2008~2017 for a while.

Some wrapper calls are also added to setup.py so that installations using
setuptools can be supported, like what is done for the autotools buids,
for release builds.